Refactor: system.py
